RDCC-178-spring-auto-wire refactor completed
