+++ src/functionalTest/java/uk/gov/hmcts/reform/professionalapi/client/ProfessionalApiClient.java
@@ -5,6 +5,7 @@ import static org.apache.commons.lang.RandomStringUtils.randomAlphabetic;
 import static org.springframework.http.HttpStatus.CREATED;
 import static org.springframework.http.HttpStatus.OK;
 import static uk.gov.hmcts.reform.professionalapi.controller.request.ContactInformationCreationRequest.aContactInformationCreationRequest;
+import static uk.gov.hmcts.reform.professionalapi.controller.request.DxAddressCreationRequest.dxAddressCreationRequest;
 import static uk.gov.hmcts.reform.professionalapi.controller.request.PbaAccountCreationRequest.aPbaPaymentAccount;
 import static uk.gov.hmcts.reform.professionalapi.controller.request.UserCreationRequest.aUserCreationRequest;
 import static uk.gov.hmcts.reform.professionalapi.utils.OrganisationFixtures.someMinimalOrganisationRequest;
@@ -78,6 +79,10 @@ public class ProfessionalApiClient {
 
         OrganisationCreationRequest organisationCreationRequest = someMinimalOrganisationRequest()
                         .name(organisationName)
+                        .sraId(randomAlphabetic(10) + "sra-id-number1")
+                        .sraRegulated(Boolean.FALSE)
+                        .companyUrl(randomAlphabetic(10) + "company-url")
+                        .companyNumber(randomAlphabetic(5) + "com")
                         .pbaAccounts(pbaAccounts)
                         .superUser(aUserCreationRequest()
                                 .firstName("some-fname")
@@ -85,7 +90,12 @@ public class ProfessionalApiClient {
                                 .email(randomAlphabetic(10) + "@somewhere.com")
                                 .build())
                         .contactInformation(Arrays.asList(aContactInformationCreationRequest()
-                                .addressLine1("addressLine1").build()))
+
+                                .addressLine1("addressLine1")
+                                .dxAddress(Arrays.asList(dxAddressCreationRequest()
+                                        .dxNumber("DX 1234567890")
+                                        .dxExchange("dxExchange").build()))
+                                .build()))
                         .build();
 
         Response response = withAuthenticatedRequest()
