+++ src/integrationTest/java/uk/gov/hmcts/reform/professionalapi/CreateMinimalOrganisationTest.java
@@ -170,8 +170,6 @@ public class CreateMinimalOrganisationTest extends Service2ServiceEnabledIntegra
         Map<String, Object> response =
                 professionalReferenceDataClient.createOrganisation(organisationCreationRequest);
 
-        //! assertThat(response.get("http_status")).isEqualTo("500");
-        //! assertThat(response.get("response_body")).isEqualTo("Error");
-
+        assertThat(response.get("http_status")).isEqualTo("500");
     }
 }
