+++ src/integrationTest/java/uk/gov/hmcts/reform/professionalapi/CreateOrganisationWithContactInformationDxAddress.java
@@ -14,7 +14,6 @@ import java.util.Map;
 import java.util.UUID;
 
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import uk.gov.hmcts.reform.professionalapi.controller.request.OrganisationCreationRequest;
@@ -315,7 +314,6 @@ public class CreateOrganisationWithContactInformationDxAddress extends Service2S
         assertThat(professionalUserRepository.findAll().size()).isEqualTo(0);
     }
 
-    @Ignore
     @Test
     public void returns_bad_request_when_contact_information_null() {
