+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/OrganisationController.java
@@ -6,13 +6,14 @@ import io.swagger.annotations.ApiResponses;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import uk.gov.hmcts.reform.professionalapi.controller.request.OrganisationCreationRequest;
 import uk.gov.hmcts.reform.professionalapi.controller.request.OrganisationCreationRequestValidator;
 import uk.gov.hmcts.reform.professionalapi.controller.response.OrganisationResponse;
-import uk.gov.hmcts.reform.professionalapi.service.OrganisationService;
+import uk.gov.hmcts.reform.professionalapi.service.impl.OrganisationServiceImpl;
 
 
 @RequestMapping(
@@ -24,16 +25,10 @@ import uk.gov.hmcts.reform.professionalapi.service.OrganisationService;
 @Slf4j
 public class OrganisationController {
 
-    private final OrganisationService organisationService;
-    private final OrganisationCreationRequestValidator validator;
-
-    public OrganisationController(
-            OrganisationService organisationService,
-            OrganisationCreationRequestValidator validator) {
-
-        this.organisationService = organisationService;
-        this.validator = validator;
-    }
+    @Autowired
+    OrganisationServiceImpl organisationService;
+    @Autowired
+    OrganisationCreationRequestValidator validator;
 
     @ApiOperation("Creates an organisation")
     @ApiResponses({
