+++ src/main/java/uk/gov/hmcts/reform/professionalapi/persistence/OrganisationRepository.java
@@ -2,8 +2,10 @@ package uk.gov.hmcts.reform.professionalapi.persistence;
 
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 import uk.gov.hmcts.reform.professionalapi.domain.Organisation;
 
+@Repository
 public interface OrganisationRepository extends JpaRepository<Organisation, UUID> {
 
     Organisation findByName(String name);
