+++ src/main/java/uk/gov/hmcts/reform/professionalapi/persistence/ProfessionalUserRepository.java
@@ -2,7 +2,9 @@ package uk.gov.hmcts.reform.professionalapi.persistence;
 
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 import uk.gov.hmcts.reform.professionalapi.domain.ProfessionalUser;
 
+@Repository
 public interface ProfessionalUserRepository extends JpaRepository<ProfessionalUser, UUID> {
 }
